// src/types/supplier.ts
export interface Supplier {
    id: number
    name: string
    image: string
    location: string
    rating: number
    reviews: number
    verified: boolean
    specialties: string[]
    description: string
    products: number
    deliveryTime: string
    minOrder: string
    featured: boolean
  }
  